+++ src/RandomPlayer.cpp
@@ -10,7 +10,7 @@ RandomPlayer::RandomPlayer(int numRows, int numCols, Board* iboard)
 {
 	id = 2;
 	distribution = uniform_int_distribution<int>(1,numCols);
-	generator.seed(chrono::system_clock::now().time_since_epoch().count());
+	generator.seed((unsigned int)chrono::system_clock::now().time_since_epoch().count());
 	dice = bind(distribution, generator);
 }
 
@@ -19,7 +19,7 @@ RandomPlayer::RandomPlayer(Board* iboard)
 {
 	id = 2;
 	distribution = uniform_int_distribution<int>(1,iboard->GetCol());
-	generator.seed(chrono::system_clock::now().time_since_epoch().count());
+	generator.seed((unsigned int)chrono::system_clock::now().time_since_epoch().count());
 	dice = bind(distribution, generator);
 }
